Split task manager UI from window controller
