+++ Sources/TaskManagerWindowController.swift
@@ -48,7 +48,7 @@ final class TaskManagerWindowController: NSWindowController, NSWindowDelegate {
 }
 
 @MainActor
-private final class CmuxTaskManagerModel: ObservableObject {
+final class CmuxTaskManagerModel: ObservableObject {
     @Published private(set) var snapshot = CmuxTaskManagerSnapshot.empty
     @Published private(set) var isRefreshing = false
     @Published private(set) var errorMessage: String?
@@ -272,262 +272,5 @@ private final class CmuxTaskManagerModel: ObservableObject {
         guard let targetSurfaceId,
               let panel = workspace.panels[targetSurfaceId] else { return }
         panel.triggerFlash(reason: .debug)
-    }
-}
-
-private struct CmuxTaskManagerView: View {
-    @ObservedObject var model: CmuxTaskManagerModel
-
-    var body: some View {
-        VStack(spacing: 0) {
-            toolbar
-            Divider()
-            summary
-            Divider()
-            tableHeader
-            Divider()
-            tableBody
-        }
-        .frame(minWidth: 820, minHeight: 480)
-        .onAppear {
-            model.start()
-        }
-        .onDisappear {
-            model.stop()
-        }
-    }
-
-    private var toolbar: some View {
-        HStack(spacing: 12) {
-            Text(String(localized: "taskManager.title", defaultValue: "Task Manager"))
-                .font(.title3.weight(.semibold))
-
-            if model.isRefreshing {
-                ProgressView()
-                    .controlSize(.small)
-                    .accessibilityLabel(String(localized: "taskManager.refreshing", defaultValue: "Refreshing"))
-            }
-
-            Spacer()
-
-            Toggle(
-                String(localized: "taskManager.showProcesses", defaultValue: "Processes"),
-                isOn: $model.includesProcesses
-            )
-            .toggleStyle(.checkbox)
-
-            Button {
-                model.refresh(force: true)
-            } label: {
-                Label(String(localized: "taskManager.refresh", defaultValue: "Refresh"), systemImage: "arrow.clockwise")
-            }
-        }
-        .padding(.horizontal, 16)
-        .padding(.vertical, 12)
-    }
-
-    private var summary: some View {
-        HStack(spacing: 24) {
-            metric(
-                title: String(localized: "taskManager.summary.cpu", defaultValue: "CPU"),
-                value: CmuxTaskManagerFormat.cpu(model.snapshot.total.cpuPercent)
-            )
-            metric(
-                title: String(localized: "taskManager.summary.memory", defaultValue: "Memory"),
-                value: CmuxTaskManagerFormat.bytes(model.snapshot.total.residentBytes)
-            )
-            metric(
-                title: String(localized: "taskManager.summary.processes", defaultValue: "Processes"),
-                value: "\(model.snapshot.total.processCount)"
-            )
-            metric(
-                title: String(localized: "taskManager.summary.updated", defaultValue: "Updated"),
-                value: model.snapshot.updatedText
-            )
-            Spacer()
-        }
-        .padding(.horizontal, 16)
-        .padding(.vertical, 10)
-    }
-
-    private func metric(title: String, value: String) -> some View {
-        VStack(alignment: .leading, spacing: 2) {
-            Text(title)
-                .font(.caption)
-                .foregroundStyle(.secondary)
-            Text(value)
-                .font(.system(.body, design: .monospaced).weight(.semibold))
-                .monospacedDigit()
-        }
-    }
-
-    private var tableHeader: some View {
-        HStack(spacing: 8) {
-            Text(String(localized: "taskManager.column.name", defaultValue: "Name"))
-                .frame(maxWidth: .infinity, alignment: .leading)
-            Text(String(localized: "taskManager.column.cpu", defaultValue: "CPU"))
-                .frame(width: 82, alignment: .trailing)
-            Text(String(localized: "taskManager.column.memory", defaultValue: "Memory"))
-                .frame(width: 96, alignment: .trailing)
-            Text(String(localized: "taskManager.column.processes", defaultValue: "Proc"))
-                .frame(width: 58, alignment: .trailing)
-        }
-        .font(.system(size: 11, weight: .semibold))
-        .foregroundStyle(.secondary)
-        .padding(.horizontal, 16)
-        .padding(.vertical, 5)
-    }
-
-    @ViewBuilder
-    private var tableBody: some View {
-        if let errorMessage = model.errorMessage {
-            CmuxTaskManagerMessageView(
-                title: String(localized: "taskManager.error.title", defaultValue: "Unable to load resource usage"),
-                detail: errorMessage
-            )
-        } else if model.snapshot.rows.isEmpty {
-            CmuxTaskManagerMessageView(
-                title: String(localized: "taskManager.empty.title", defaultValue: "No resource usage"),
-                detail: String(localized: "taskManager.empty.detail", defaultValue: "Open a workspace, terminal, or browser surface to see it here.")
-            )
-        } else {
-            ScrollView {
-                LazyVStack(spacing: 0) {
-                    ForEach(model.snapshot.rows) { row in
-                        CmuxTaskManagerRowView(
-                            row: row,
-                            onViewWorkspace: {
-                                model.viewWorkspace(for: row)
-                            },
-                            onViewTerminal: {
-                                model.viewTerminal(for: row)
-                            },
-                            onKillProcess: {
-                                model.killProcess(for: row)
-                            },
-                            onActivate: {
-                                model.viewBestTarget(for: row)
-                            }
-                        )
-                        Divider()
-                            .padding(.leading, 16)
-                    }
-                }
-            }
-        }
-    }
-}
-
-private struct CmuxTaskManagerMessageView: View {
-    let title: String
-    let detail: String
-
-    var body: some View {
-        VStack(spacing: 8) {
-            Text(title)
-                .font(.headline)
-            Text(detail)
-                .font(.callout)
-                .foregroundStyle(.secondary)
-                .multilineTextAlignment(.center)
-                .textSelection(.enabled)
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-        .padding(32)
-    }
-}
-
-private struct CmuxTaskManagerRowView: View {
-    let row: CmuxTaskManagerRow
-    let onViewWorkspace: () -> Void
-    let onViewTerminal: () -> Void
-    let onKillProcess: () -> Void
-    let onActivate: () -> Void
-
-    var body: some View {
-        HStack(spacing: 8) {
-            HStack(spacing: 5) {
-                Color.clear
-                    .frame(width: CGFloat(row.level) * 14)
-                rowIcon
-                VStack(alignment: .leading, spacing: 0) {
-                    Text(row.title)
-                        .font(.system(size: 12.5))
-                        .lineLimit(1)
-                    if !row.detail.isEmpty {
-                        Text(row.detail)
-                            .font(.system(size: 11))
-                            .foregroundStyle(.secondary)
-                            .lineLimit(1)
-                    }
-                }
-            }
-            .frame(maxWidth: .infinity, alignment: .leading)
-
-            Text(CmuxTaskManagerFormat.cpu(row.resources.cpuPercent))
-                .frame(width: 82, alignment: .trailing)
-            Text(CmuxTaskManagerFormat.bytes(row.resources.residentBytes))
-                .frame(width: 96, alignment: .trailing)
-            Text("\(row.resources.processCount)")
-                .frame(width: 58, alignment: .trailing)
-        }
-        .font(.system(size: 12.5, design: .default))
-        .monospacedDigit()
-        .padding(.horizontal, 16)
-        .padding(.vertical, 3)
-        .opacity(row.isDimmed ? 0.68 : 1)
-        .contentShape(Rectangle())
-        .onTapGesture {
-            onActivate()
-        }
-        .contextMenu {
-            if row.canViewWorkspace {
-                Button {
-                    onViewWorkspace()
-                } label: {
-                    Label(
-                        String(localized: "taskManager.contextMenu.viewWorkspace", defaultValue: "View Workspace"),
-                        systemImage: "rectangle.stack"
-                    )
-                }
-            }
-            if row.canViewTerminal {
-                Button {
-                    onViewTerminal()
-                } label: {
-                    Label(
-                        String(localized: "taskManager.contextMenu.viewTerminal", defaultValue: "View Terminal"),
-                        systemImage: "terminal"
-                    )
-                }
-            }
-            if row.canKillProcess {
-                Divider()
-                Button {
-                    onKillProcess()
-                } label: {
-                    Label(
-                        String(localized: "taskManager.contextMenu.killProcess", defaultValue: "Kill Process..."),
-                        systemImage: "xmark.octagon"
-                    )
-                }
-            }
-        }
-    }
-
-    @ViewBuilder
-    private var rowIcon: some View {
-        if let agentAssetName = row.agentAssetName {
-            Image(agentAssetName)
-                .resizable()
-                .interpolation(.high)
-                .aspectRatio(contentMode: .fit)
-                .frame(width: 14, height: 14)
-        } else {
-            Image(systemName: row.kind.systemImage)
-                .foregroundStyle(row.kind.tint)
-                .font(.system(size: 12))
-                .frame(width: 14)
-        }
     }
 }
